AX: With ENABLE(AX_THREAD_TEXT_APIS), we can crash when using cached selection text marker ranges constructed from VisiblePositionRange / VisibleSelection https://bugs.webkit.org/show_bug.cgi?id=286710 rdar://143849705
Reviewed by NOBODY (OOPS!).

Prior to this commit, we would call `AXIsolatedTree::setSelectedTextMarkerRange` with an AXTextMarkerRange constructed
from a VisiblePositionRange. This is a problem, since offsets in VisiblePositionRange are pre-whitespace collapse,
while the text runs we cache off the main-thread are post-whitespace collapse, meaning the offsets would reguarly point
to the wrong character, or completely out-of-bounds of the text runs, causing crashes.

With this commit, for ENABLE(AX_THREAD_TEXT_APIS), we instead construct the selection AXTextMarkerRange using
VisiblePosition::inlineBoxAndOffset(), which does all the necessary work to turn the VisiblePosition into an offset
that points into some inline box, which is the basis for the text runs we cache off the main-thread.

This fixes mac/dynamic-selection.html in ENABLE(AX_THREAD_TEXT_APIS), and unblocks the ability to fix more tests which
now fail for other reasons that are not selection related.

* Source/WebCore/accessibility/AXTextMarker.cpp:
(WebCore::AXTextMarkerRange::AXTextMarkerRange):
* Source/WebCore/accessibility/AXTextMarker.h:
* Source/WebCore/accessibility/isolatedtree/AXIsolatedTree.cpp:
(WebCore::AXIsolatedTree::create):
* Source/WebCore/accessibility/mac/AXObjectCacheMac.mm:
(WebCore::AXObjectCache::onSelectedTextChanged):